+++ src/sdk/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Web branding via `IBrandingManager` with support for getting and setting themes and site chrome (header/footer/navigation) [jansenbe - Bert Jansen]
 - Added admin library support for enumerating all webs in a given site collection [jansenbe - Bert Jansen]
 - EventHub for system notifications - retryevent is the first event being handled, allowing to understand when code get's throttled [jansenbe - Bert Jansen]
+- Properties on `PnPContext` to enable adding context. These properties are passed along in case of an EventHub notification [jansenbe - Bert Jansen]
 
 ### Changed